Translationfix: makes a wide char copy of the scripts + bugfixes
While checking the txt files and writing the updated files, a widechar copy is saved in musume\scripts.
This copy is then included when musume is copied to the game after the txt file check is complete.
The unnamed shop files are now generated only for the widechar copy.

Since converting to widechar takes twice as long as the rest of the script combined, an on/off switch has been added.
$make_widechar_copy can be set to 0 to skip the widechar version, which could be useful in certain cases, like when fixing script bugs.
Default is 1, which mean the widechar version will be made.

The screen flashing code is now correctly inserted again to use the global setting set by the TL forewords