Bugfix: stat improvements in training will stack instead of use the last one
The problem is that most training activities had code expecting one training session to contain only one event.
Due to this expectation, it is assumed that the already gained stats for that day's training are zero and safe to overwrite.

As a result, if activity X is done twice, only the second time counts when stats are added in the end of the training session.

The fix adds the old value to the new one, ensuring that the stat gain is for all activities combined, not just the last one.

Issue pointed out by SexyKisune.