Translationfix: the level up info in translation status is now always written in the same order
The problem was that I requested all stats to be printed, but not which order. It was consistent in order and it looked ok.
Turns out that it is consistent as long as it is run on the same computer, but not when different computers ran the same code.

Now the order is set to be body, mind, int, which is the same order as they appear ingame.