Multiple minor corrections. All files now run through all 4 phases
There are a bit of an issue with 3 files. Looks like they might require a bit of editing in order to work as intended.
The reason is corrections, which the conversion script can't handle. Luckily it's just one line in each file.

The files are:
status/karin_kenkou.txt
schedule/setsumei.txt
day/zakuro4.txt